refactor: split RegistrationConfirmationPage into RegistrationConfirmationForm and useRegistrationConfirmation
