+++ src/app/auth/registration-confirmation/registration-confirmation-form/RegistrationConfirmationForm.tsx
@@ -1,0 +1,27 @@
+'use client'
+import { Button } from '@/src/components/button/Button'
+import { Typography } from '@/src/components/typography/Typography'
+import { useRegistrationConfirmation } from '@/src/hooks/useRegistrationConfirmation'
+import Image from 'next/image'
+import Link from 'next/link'
+
+import s from '@/src/app/auth/registration-confirmation/registration-confirmation-form/email-confirmed.module.scss'
+
+export const RegistrationConfirmationForm = () => {
+  useRegistrationConfirmation()
+
+  return (
+    <div className={s.container}>
+      <Typography as={'h1'} option={'h1'}>
+        {'Congratulations!'}
+      </Typography>
+      <Typography as={'h2'} option={'regular_text16'}>
+        {'Your email has been confirmed'}
+      </Typography>
+      <Button as={Link} href={'/auth/login'} variant={'primary'}>
+        {'Sing In'}
+      </Button>
+      <Image alt={''} height={300} src={'/image/bro.svg'} width={432} />
+    </div>
+  )
+}
