+++ src/app/auth/registration-confirmation/registration-confirmation-form/email-confirmed.module.scss
