+++ src/app/auth/registration/registration-form/registration.module.scss
