+++ src/hooks/useRegistrationConfirmation.ts
@@ -1,0 +1,31 @@
+'use client'
+import { useEffect } from 'react'
+
+import { useRegistrationConfirmationMutation } from '@/src/store/services/authApi'
+import { CustomerError } from '@/src/store/services/types'
+import { useRouter, useSearchParams } from 'next/navigation'
+
+export const useRegistrationConfirmation = () => {
+  const searchParams = useSearchParams()
+  const router = useRouter()
+  const [confirmRegistration] = useRegistrationConfirmationMutation()
+
+  useEffect(() => {
+    ;(async () => {
+      try {
+        // Получаем параметры через useSearchParams
+        const code = searchParams.get('code') as string
+
+        await confirmRegistration({ confirmationCode: code }).unwrap()
+      } catch (err) {
+        const error = err as CustomerError
+        const errorMessage = error.data?.messages[0]?.message
+
+        // Проверяем сообщение об ошибке и выполняем перенаправление при необходимости
+        if (errorMessage === 'Confirmation code is invalid') {
+          router.push('/auth/registration-email-resending')
+        }
+      }
+    })()
+  }, [searchParams, confirmRegistration, router])
+}
